Fixes linked image missing alternative text
Images that are linked and that otherwise have no text equivalent from a missing, or empty, alt attribute are not WCAG compliant.
Ultimately this creates a situation where the purpose of the link is unclear when announced by a screen reader.
There are several ways to fix that, but adding text to the image's alt attribute to describe what the link does is going to be the easiest.