Fixes missing alt attribute on Spirit Island game link card
The alt attribute is required on images, even if you wanted to leave them empty.
This attribute allows you to describe the image, and convey any text that might be on the image if important.
Leaving the attribute empty means the image exists purely for decorative purposes (and would be perfectly fine in this context, fwiw)